requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.0.10
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
django-modelcluster==6.4
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
l18n==2021.3
laces==0.1.1
mysqlclient==2.2.6
openpyxl==3.1.5
pillow==10.4.0
pillow_heif==0.21.0
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.12.2
urllib3==2.3.0
wagtail==6.2.3
wagtailcodeblock==1.29.0.2
whitenoise==6.8.2
Willow==1.9.0
